Add ActionSettingInput
